EventBComponent's order of supertypes switched back.
Order of the classifiers used as supertypes for EventBComponent is
switched back to AbstractExtension (first), Component (second) to fix
code generation problems and persistence problems with extended
Machines.

Additionally, the default extension id for EventBComponent is set in a
constructor to a default constant (temporary solution for resetting the
id constant defined in the AbstractExtension superclass).