Modification to metamodel.
The ports and FMU internal variables are now subclassed from the
abstract variable that includes type and causality. The metamodel also
adds time attribute (of type double) to the component diagram class
(diagram root element) to hold the global simulation time.

All model source, diagram and UI are updated to support the changes in
the metamodel.

NOTE: new metamodel is not compatible with the old one, so old model
files will not work with the new code.